auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "TMS auth-service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.5",
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.2.2",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.9.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-dotenv>=0.5.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
]
